Refactor the BaseLlcMotion class to have subclasses with and without crawl.
